pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tfadm"
version = "0.11.0"
requires-python = ">=3.7"
description = "Generates and modifies terraform code in JSON syntax."
readme = "README.md"
authors = [
  {name = "Nuno Cardoso"}
]
license = {file = "LICENSE"}
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: System Administrators",
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Topic :: System :: Systems Administration",
]
dependencies = [
  "click>=8.0",
  "Jinja2>=3.0.0",
  "parse>=1.19.0",
  "python-slugify>=7.0.0",
  "pyyaml>=6.0",
]

[project.urls]
"Homepage" = "https://github.com/nuncard/tfadm"
"Bug Tracker" = "https://github.com/nuncard/tfadm/issues"

[project.scripts]
  tfadm = "tfadm.__main__:main"
